% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_new.R
\name{filter_tx_new}
\alias{filter_tx_new}
\title{Filter TX_NEW Patients}
\usage{
filter_tx_new(linelist, start_date, end_date, transfer_column)
}
\arguments{
\item{linelist}{A data frame containing the patient data.}

\item{start_date}{The start date for the filtering period.}

\item{end_date}{The end date for the filtering period.}

\item{transfer_column}{The name of the column indicating if the patient was transferred in.}
}
\value{
A filtered data frame with new patients (TX_NEW).
}
\description{
Filters the patients based on their ART start date and transfer status.
}
