% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(filepath)
}
\arguments{
\item{filepath}{The path to the CSV file.}
}
\value{
A dataframe containing the loaded data.
}
\description{
This function loads data from a CSV file and returns it as a dataframe.
}
\examples{
linelist <- load_data("path/to/tll_6_6_2024.csv")
}
